Fixes issue 5
This update Fixes #5 . Now, we may have as many arbitrary  `activation` functions we may want while designing the Neural Network system. Also, doing this helped improve the code's readability. The previous implementation relied on generic classes too much, and this affected the elegance of the code. Now the design was refined to avoid generic classes and provide the functionality of adding arbitrary activation functions to each neurons.