Fixed a lot of bugs and issues.
The current upload closes #12 and Fixes $16. Though the original demands of #16 aren't met, the issue was solved by an alternate solution. #16 clearly expects us to create error handling code, that encounters values going out of bounds (overflow and underflow). But the requirement of such handling code was high initially because of the numerous bugs in the early version. There were too many bugs that it was hardly useful for any training purpose. 

Looks like these issues are solved.  Not the high priority is to solve the issue no. #9. This expects major changes in the source. #6 must also be considered, but as of now it is not at the top of the list.